+++ FrontEnd/App.js
@@ -5,9 +5,9 @@ import { Provider, connect } from 'react-redux';
 import { createStore } from 'redux';
 import loginReducer from './src/reducers/loginReducer';
 import teacherNavigationReducer from "./src/reducers/teacherNavigationReducer";
-import LoginScreen from "./src/components/LoginScreen";
+import LoginScreen from "./src/components/general/LoginScreen";
 import {combineReducers, install } from 'redux-loop';
-import Navigation from "./src/components/Navigation";
+import Navigation from "./src/components/general/Navigation";
 import courseNavigationReducer from "./src/reducers/courseNavigationReducer";
 import localStorage from "./utils/localStorageMock";
 import {loadUserData, setScreenWidthIsMobile} from "./src/actions/loginActions";
