+++ FrontEnd/src/components/general/identifiers.js
@@ -1,0 +1,14 @@
+export const SUBJECT_OVERVIEW_IDENTIFIER = "SUBJECT_OVERVIEW";
+export const TASK_OVERVIEW_IDENTIFIER = "TASK_OVERVIEW";
+export const SCHEDULE_IDENTIFIER = "SCHEDULE";
+export const COMMUNICATION_DASHBOARD_IDENTIFIER = "COMMUNICATION_DASHBOARD";
+export const COURSES_FOR_SUBJECT_IDENTIFIER = "COURSES_FOR_SUBJECT";
+export const COURSE_VIEW_IDENTIFIER = "COURSE_VIEW";
+export const TASKS_FOR_COURSE_IDENTIFIER = "TASKS_FOR_COURSE";
+export const STUDENTS_FOR_COURSE_IDENTIFIER = "STUDENTS_FOR_COURSE";
+export const ACTIONS_FOR_COURSE_IDENTIFIER = "ACTIONS_FOR_COURSE";
+export const CLASS_ADMINISTRATION_IDENTIFIER = "CLASS_ADMINISTRATION_IDENTIFIER";
+export const COURSE_ADMINISTRATION_IDENTIFIER = "COURSE_ADMINISTRATION_IDENTIFIER";
+export const SUBJECT_ADMINISTRATION_IDENTIFIER = "SUBJECT_ADMINISTRATION_IDENTIFIER";
+export const USER_ADMINISTRATION_IDENTIFIER = "USER_ADMINISTRATION_IDENTIFIER";
+export const GRADES_OVERVIEW_IDENTIFIER = "GRADES_OVERVIEW";
