+++ FrontEnd/src/components/teacher/CoursesForSubject.js
@@ -6,9 +6,9 @@ import generalStyles from "../../styles/GeneralStyles";
 import Button from "@material-ui/core/Button";
 import {switchContent} from "../../actions/teacherNavigationActions";
 import {setCourse, unselectCourse} from "../../actions/subjectSelectActions";
-import {COURSE_VIEW_IDENTIFIER} from "./TeacherTabs";
 import Typography from "@material-ui/core/Typography";
 import Box from "@material-ui/core/Box";
+import {COURSE_VIEW_IDENTIFIER} from "../general/identifiers";
 
 function CoursesForSubject(props){
