+++ FrontEnd/src/reducers/teacherNavigationReducer.js
@@ -1,6 +1,6 @@
 import { SWITCH_TO_CONTENT, SHOW_EVALUATE_TASK_PAGE } from '../actions/teacherNavigationActions';
-import {SUBJECT_OVERVIEW_IDENTIFIER} from '../components/teacher/TeacherTabs';
 import { ERROR_CONTENT_IDENTIFIER } from "../actions/errorActions";
+import {SUBJECT_OVERVIEW_IDENTIFIER} from "../components/general/identifiers";
 
 //Active Content in Main Panel defaults zu diesem Wert. Für Testzwecke, kann die zu testende Seite hier eingetragen werden
 const DEFAULT_STATE = {activeContent: SUBJECT_OVERVIEW_IDENTIFIER}
